requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
arm-mango==1.5.1      
astunparse==1.6.3
attrs==23.2.0
autograd==1.6.2
autograd-gamma==0.5.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chinese_calendar==1.9.1
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
contourpy==1.2.1
cryptography==44.0.0
cycler==0.12.1
dask==2024.10.0
ephem==4.1.5
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fastjsonschema==2.20.0
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.51.0
formulaic==1.0.1
fsspec==2024.10.0
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
holidays==0.48
huggingface-hub==0.29.3
idna==3.7
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==8.5.0
interface-meta==1.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.0
keras==3.9.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lifelines==0.28.0
locket==1.0.0
LunarCalendar==0.0.9
lxml==5.2.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mysql-connector-python==8.3.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.2
openpyxl-image-loader==1.0.5
opt_einsum==3.4.0
optree==0.14.1
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
partd==1.4.2
patsy==0.5.6
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.3.0
plyer==2.1.0
protobuf==5.29.3
psycopg2==2.9.9
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.1
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
pypiwin32==223
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytups==0.88.1
pytz==2023.4
pywin32==308
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
rich==13.9.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
selenium==4.22.0
shapely==2.0.7
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
statsmodels==0.14.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.4.0
tifffile==2025.3.13
tokenizers==0.21.1
toolz==1.0.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.15.2
tqdm==4.66.6
transformers==4.49.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.8.0
Werkzeug==3.1.3
win10toast==0.9
wrapt==1.16.0
wsproto==1.2.0
xgboost==3.0.0
xlrd==2.0.1
XlsxWriter==3.2.0
zipp==3.20.2
